+++ attractions/checkbox/checkbox-group.svelte
@@ -51,9 +51,13 @@
    * @type {number}
    */
   export let max = Infinity;
+  /**
+   * The text displayed in the tooltip when hovering over the checkboxes
+   *   after the maximum allowed selection has been reached.
+   */
+  export let maxReachedTooltip = `Can only select ${max} value${s(max)}.`;
 
   $: currentChecked = items.reduce((acc, elt) => acc + elt.checked, 0);
-  $: maxReachedTooltip = `Can only select ${max} value${s(max)}.`;
 
   if (!items || items.length === 0) {
     console.error('Must have at least one item in the checkbox group');