+++ attractions/svelte.config.js
@@ -1,15 +1,15 @@
 const path = require('path');
 const autoPreprocess = require('svelte-preprocess');
 
+const makeAttractionsImporter = require('./importer.js');
+
 module.exports = {
   preprocess: [
     autoPreprocess({
       scss: {
-        importer(url, _prev) {
-          if (url === 'node_modules/attractions/_variables') {
-            return { file: path.join(__dirname, '_variables.scss') };
-          }
-        },
+        importer: makeAttractionsImporter({
+          themeFile: path.join(__dirname, '_variables.scss'),
+        }),
         renderSync: true,
       },
     }),