+++ docs/src/routes/docs/_layout.svelte
@@ -60,6 +60,10 @@
           segment: 'components/badge',
         },
         {
+          title: 'Breadcrumbs',
+          segment: 'components/breadcrumbs',
+        },
+        {
           title: 'Button',
           segment: 'components/button',
         },