+++ CHANGELOG.md
@@ -4,6 +4,12 @@ All notable changes to this project will be documented here.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/) and this project adheres to [Semantic Versioning](https://semver.org/).
 
+## [Unreleased]
+
+### Added
+
+- `maxReachedTooltip` prop for the `CheckboxGroup` component.
+
 ## [3.3.0] - 2021-05-28
 
 This release contains no new features but fixes a Sass deprecation. You'll have to bump your `sass` dependency to `>= 1.33.0` in order to upgrade to this version.