+++ CHANGELOG.md
@@ -4,7 +4,11 @@ All notable changes to this project will be documented here.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/) and this project adheres to [Semantic Versioning](https://semver.org/).
 
-## Unreleased
+## [Unreleased]
+
+### Added
+
+- The `hours`, `minutes` and `seconds` props of the `TimePicker` component allow you to specify what values the user can choose from.
 
 ### Fixed