+++ CHANGELOG.md
@@ -11,6 +11,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/) and this 
 - The `hours`, `minutes` and `seconds` props of the `TimePicker` component allow you to specify what values the user can choose from.
 - `disabledDates` prop for the `Calendar` component to disable specific dates (or ranges of dates).
 
+### Fixed
+
+- _[For contributors on Windows]_ The scripts in `package.json` should now work on CMD as well.
+
 ## [3.3.0] - 2021-05-28
 
 This release contains no new features but fixes a Sass deprecation. You'll have to bump your `sass` dependency to `>= 1.33.0` in order to upgrade to this version.