+++ CHANGELOG.md
@@ -4,7 +4,7 @@ All notable changes to this project will be documented here.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/) and this project adheres to [Semantic Versioning](https://semver.org/).
 
-## [Unreleased]
+## [3.4.0] - 2021-09-12
 
 ### Added
 
@@ -178,7 +178,8 @@ Minor documentation and bug fixes.
 First stable release with proper documentation.
 Previous, undocumented, releases can be found in [the releases section](https://github.com/illright/attractions/releases).
 
-[unreleased]: https://github.com/illright/attractions/compare/v3.3.0...HEAD
+[unreleased]: https://github.com/illright/attractions/compare/v3.4.0...HEAD
+[3.4.0]: https://github.com/illright/attractions/releases/tag/v3.4.0
 [3.3.0]: https://github.com/illright/attractions/releases/tag/v3.3.0
 [3.2.1]: https://github.com/illright/attractions/releases/tag/v3.2.1
 [3.2.0]: https://github.com/illright/attractions/releases/tag/v3.2.0